- fixed MouseEvents for different resolutions

+++ src/de/gemo/engine/events/mouse/MouseDragEvent.java
@@ -2,21 +2,21 @@ package de.gemo.engine.events.mouse;
 
 public class MouseDragEvent extends AbstractMouseClickEvent {
 
-    private final int difX, difY;
+    private final float difX, difY;
     private final boolean moved;
 
-    public MouseDragEvent(int x, int y, int difX, int difY, MouseButton button) {
+    public MouseDragEvent(int x, int y, float difX, float difY, MouseButton button) {
         super(x, y, button, 3);
         this.difX = difX;
         this.difY = difY;
         this.moved = (difX != 0 || difY != 0);
     }
 
-    public int getDifX() {
+    public float getDifX() {
         return difX;
     }
 
-    public int getDifY() {
+    public float getDifY() {
         return difY;
     }
