+++ src/de/gemo/engine/events/mouse/MouseMoveEvent.java
@@ -2,19 +2,19 @@ package de.gemo.engine.events.mouse;
 
 public class MouseMoveEvent extends AbstractMouseEvent {
 
-    private final int difX, difY;
+    private final float difX, difY;
 
-    public MouseMoveEvent(int x, int y, int difX, int difY) {
+    public MouseMoveEvent(int x, int y, float difX, float difY) {
         super(x, y, 0);
         this.difX = difX;
         this.difY = difY;
     }
 
-    public int getDifX() {
+    public float getDifX() {
         return difX;
     }
 
-    public int getDifY() {
+    public float getDifY() {
         return difY;
     }
 }
